+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.6.0
+- Can select if chat, emotes, or actions can trigger the clicker independently
+- Crawl into bed with someone activity on arms
+
 # 0.5.0
 - Owners option subpage
 - Owner rework - Allows more authority control for the user to restrict access to the authority but still add / remove others